+++ packages/terra-date-picker/src/DateInput.jsx
@@ -5,6 +5,7 @@ import Button from 'terra-button';
 import IconCalendar from 'terra-icon/lib/icon/IconCalendar';
 import Input from 'terra-form/lib/Input';
 import DateUtil from './DateUtil';
+import styles from './DatePicker.scss';
 
 const Icon = <IconCalendar />;
 
@@ -126,18 +127,18 @@ class DatePickerInput extends React.Component {
     const dateValue = DateUtil.convertToISO8601(value, DateUtil.getFormatByLocale(this.context.intl.locale));
 
     return (
-      (<div className="terra-DatePicker-customInput">
+      (<div className={styles['custom-input']}>
         <input
           // Create a hidden input for storing the name and value attributes to use when submitting the form.
           // The data stored in the value attribute will be the visible date in the date input but in ISO 8601 format.
-          className="terra-hidden-date-input"
+          data-class="hidden-date-input"
           type="hidden"
           name={name}
           value={dateValue}
         />
         <Input
           {...additionalInputProps}
-          className="terra-DatePicker-input"
+          className={styles.input}
           type="text"
           name={'terra-date-'.concat(name)}
           value={value}
@@ -145,7 +146,7 @@ class DatePickerInput extends React.Component {
           placeholder={placeholder}
         />
         <Button
-          className="terra-DatePicker-button"
+          className={styles.button}
           onClick={this.handleOnButtonClick}
           onKeyDown={this.handleOnKeyDown}
           icon={Icon}