+++ packages/terra-grid/tests/nightwatch/components/GridResponsive.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Grid from '../../../lib/Grid';
 
 const GridResponsive = () => (
-  <Grid>
+  <Grid id="grid">
     <Grid.Row>
       <Grid.Column tiny={12} small={8} medium={4} large={6}>
         <div className="terra-Grid-example">First Column content</div>