+++ packages/terra-site/src/TestLinks.jsx
@@ -1,10 +1,11 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import React from 'react';
 import { Link } from 'react-router';
+import styles from './site.scss';
 
 const TestLinks = () => (
-  <div>
-    <ul className="site-nav">
+  <div className={styles.app} >
+    <ul className={styles['site-nav']}>
       <li><Link to="/tests/alert-tests">Alert Tests</Link></li>
       <li><Link to="/tests/arrange-tests">Arrange Tests</Link></li>
       <li><Link to="/tests/badge-tests">Badge Tests</Link></li>
@@ -31,6 +32,7 @@ const TestLinks = () => (
       <li><Link to="/tests/popup-tests">Popup Tests</Link></li>
       <li><Link to="/tests/profile-image-tests">Profile Image Tests</Link></li>
       <li><Link to="/tests/progress-bar-tests">ProgressBar Tests</Link></li>
+      <li><Link to="/tests/props-table-tests">Props Table Tests</Link></li>
       <li><Link to="/tests/responsive-element-tests">ResponsiveElement Tests</Link></li>
       <li><Link to="/tests/search-field-tests">Search Field</Link></li>
       <li><Link to="/tests/single-select-list-tests">SingleSelectList Tests</Link></li>