+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,8 +3,19 @@ Changelog
 
 Unreleased
 ----------
+
+1.4.0 - (August 1, 2017)
+------------------
+### Changed
+* Minor version bump
+
+1.3.0 - (August 1, 2017)
+------------------
 ### Added
 * Added requestFocus and releaseFocus props to support requesting and releasing focus.
+
+### Changed
+* Converted component to use CSS modules
 
 1.2.0 - (July 26, 2017)
 ------------------