+++ packages/terra-form/CHANGELOG.md
@@ -3,11 +3,14 @@ Changelog
 
 Unreleased
 ----------
+
+1.2.0 - (August 1, 2017)
+------------------
 ### Changed
+* Converted package to use CSS Modules
 * Update Field documentation
 * Change NumberField defaultValue propType to accept string or numeric values
 * Removed red ring for required fields.
-
 
 1.1.0 - (July 18, 2017)
 ------------------