+++ packages/terra-modal-manager/CHANGELOG.md
@@ -3,13 +3,25 @@ Changelog
 
 Unreleased
 ----------
+
+1.4.0 - (August 1, 2017)
+------------------
+### Changed
+* Minor version bump
+
+1.3.0 - (August 1, 2017)
+------------------
 ### Added
 * Added gainFocus and loseFocus actions to support trapping and releasing focus in the modal manager.
+
+### Changed
+* Converted component to use CSS modules
 
 1.2.0 - (July 26, 2017)
 ------------------
 ### Changed
 * Updated spacing around react-docgen comments
+* Update nightwatch tests for Slide Group conversion to CSS modules
 
 1.1.0 - (July 18, 2017)
 ------------------