+++ packages/terra-overlay/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+1.2.0 - (August 1, 2017)
+------------------
+### Changed
+* Updated example code in overlay documentation
+
 1.1.0 - (July 26, 2017)
 ------------------
 ### Changed