+++ packages/terra-profile-image/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+1.1.0 - (August 1, 2017)
+------------------
+### Changed
+* Converted component to use CSS modules
+
 1.0.0 - (July 18, 2017)
 ------------------
 Initial stable release