+++ packages/terra-site/CHANGELOG.md
@@ -3,13 +3,24 @@ Changelog
 
 Unreleased
 ----------
+
+1.4.0 - (August 1, 2017)
+------------------
+### Changed
+* Minor version bump
+
+1.3.0 - (August 1, 2017)
+------------------
 ### Added
 * Added new example to display date picker and popup inside modal.
+* Updated PropsTable rendering to include component title for packages that contain multiple components.
 
 1.2.0 - (July 26, 2017)
 ------------------
 ### Changed
+* Converted component to use CSS modules
 * Updated Search Field example to demonstrate onInvalidSearch.
+* Add Model Manager Test Link
 
 1.1.0 - (July 18, 2017)
 ------------------