+++ packages/terra-slide-panel/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+1.2.0 - (August 1, 2017)
+------------------
+### Changed
+* Converted component to use CSS modules
+
 1.1.0 - (July 18, 2017)
 ------------------
 ### Changed